requirements.txt
# Core
langchain>=0.1.0
langchain-community
langchain-openai
openai
tiktoken

# PDF parsing
PyMuPDF

# Embeddings & vector store
faiss-cpu

# Web UI
streamlit
python-dotenv

charset-normalizer<4  # some compatibility issues with requests